Users for git review --reviewers from file in repo
We use an alias (git_review, from boot/bash_profile) to add a number of default reviewers to
each gerrit "push". This is not flexible enough. New employees are not
automatically added, the alias is a bash alias so not overrideable for
different repos, etc.

Instead, create an alias like this:

git config --global alias.rev \!"if [ -s '.gitreviewers' ]; then cat
.gitreviewers | xargs git review --reviewers; else git review; fi"

and use `git rev` from now on. This will look at a file called
.gitreviewers in the root of the repository, containing email addresses
for reviewers, one per line. If no such file exists, it falls back to
vanilla git review.

Change-Id: I9b8acaeccc240887a89a538debf0b57eb270137b
Reviewed-on: https://gerrit.ecg.so/40201
Tested-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Peter Troshin <ptroshin@ebay.com>
Reviewed-by: Jasper Van Veghel <jvanveghel@ebay.com>