KCA-17669 ReplyTS AMQ connectors do not recover from AMQ outages
Keep retrying forever to reconnect to AMQ. Otherwise, a short outage
may require a manual restart of the app. The timeout option seems
really small, but it only kicks in once a broker is down. We need to
shed load in that case quickly, because each AMQ publish op is using a
processing thread.

Change-Id: If4ec337481eb338bed0ae6c410c680bf9ead25d7
Reviewed-on: https://gerrit.ecg.so/40615
Tested-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Shiying Chen <yvonchen@ebay.com>
Reviewed-by: Ryan Prayogo <rprayogo@ebay.ca>
Reviewed-by: Gerrit-At-Gitmaster001