Change push service configuration to use url instead of host:port
We migrated our push service to a cloud hosted model and thus need to
have the host header and endpoint of the requests adhere to certain
restrictions. Therefore we replaced the host and port based configuration
with an url based one. Also the defaults of the configurations are changed
to match the new model.

Change-Id: I959e6a2e81091291946a523bfcaa1fff194c7536
Reviewed-on: https://gerrit.ecg.so/42525
Tested-by: Gerrit-At-Gitmaster001
Reviewed-by: Geert Gerritsen <ggerritsen@ebay.com>