Fix error handling in the controller
Fix by @smoczarski, ported to comaas. Fix a problem that on exception
handling in the controller the writer will be requested by Spring. This
is not allowed when the output stream was requested before (what seems
to be the case). This fix will only use the response, check if the
response can be reseted and use only then request the writer.

Change-Id: I4d9dcb0361179578c2a31a49d7cc0094e5beb41e
Reviewed-on: https://gerrit.ecg.so/42598
Tested-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Peter Troshin <ptroshin@ebay.com>
Reviewed-by: Geert Gerritsen <ggerritsen@ebay.com>
Reviewed-by: Gerrit-At-Gitmaster001