Service discovery based on tags
This patch adds a filter to services discovered through Consul. Any
Consul service needs to have a tag set with the same value as the tenant
ID it is intended for. Shared services need to have all tenant tags set,
unshared services just the tag of the tenant it is intended for.

Change-Id: I6e1b455328720c273718eb9b359eee6de89f39e8
Reviewed-on: https://gerrit.ecg.so/43022
Tested-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Geert Gerritsen <ggerritsen@ebay.com>
Reviewed-by: Jasper Van Veghel <jvanveghel@ebay.com>
Reviewed-by: Gerrit-At-Gitmaster001