Set logDir to known directory
Before, logging to file did not work due to the fact that bin/comaas
automatically sets the logDir to $BASEDIR/logs. This is a directory
inside the untarred package structure, so we cannot set permissions on
it. The logs did end up on disk because they are picked up from STDOUT
as well.
We now also want to have access logging to file, so this needed to be
fixed. I've set the logDir to /opt/replyts/logs, which should be
available and writable on all environments.

Change-Id: Iff8920faccd02186047fa1f49b502977505e04db
Reviewed-on: https://gerrit.ecg.so/50766
Tested-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Jeroen Smolenaers <jsmolenaers@ebay.com>
Reviewed-by: Remmelt Pit <rpit@ebay.com>