Rolled back to the old way of doing ES searches
This partially reverts commit 32fb6bccb3, the GTUK-1A code merge.

There are three ways to search in ES:
1) Query
2) Filtered query
3) Filtered filter

The original method used before the GTUK merge was 2) Filtered query.
This performs well on large datasets. The merge gave the choice between
1) and 3), where 1) Query performs badly on large datasets. 3) Seems to
be performing OK but does not support wildcard searches.

Now rolling back to 2), let's see if this is fast enough for GTUK's to
be merged velocity filter as well.

Change-Id: I32e90b7029a96410cdf42eafe808c953e0ea8d6d
Reviewed-on: https://gerrit.ecg.so/55226
Tested-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Remmelt Pit <rpit@ebay.com>