Fixes all the kjca VolumeFilter* flickering tests
The VolumeFilter was missing the fact that SharedBrain#markSeen
method is asynchronous, so if even a minor delay between submitting
an email into the "seen" topic and processing the hazelcast message
occurs, then processor#count() doesn't count the "current" attempt.
Instead we can freely add the current attempt (since it's happening
at the moment) to the processor#count() and mark an email as seen
afterwards.

Change-Id: Id2a1f774a352d1f44c53c647d38a177f1406c92d
Reviewed-on: https://gerrit.ecg.so/56983
Tested-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Peter Troshin <ptroshin@ebay.com>
Reviewed-by: Jasper Van Veghel <jvanveghel@ebay.com>
Reviewed-by: Gerrit-At-Gitmaster001