Removed redundant property ssl protocol
We were confused with the 'replyts.ssl.protocol' property so I spent
some time figuring out what it does. First, it's misnamed, and should be
called nextProtocol (after SSL) instead. Second, it should never be set
to anything except the default, so I changed it into a constant to avoid
future confusion.

Also found that the stopTimeout was always set to 5000ms... removed this
so we can actually set it with the property.

Change-Id: I87c85699510ae896707932b9d3c136432b817c12
Reviewed-on: https://gerrit.ecg.so/59223
Tested-by: Gerrit-At-Gitmaster001
Reviewed-by: Remmelt Pit <rpit@ebay.com>