COMAAS-280: Register Comaas with Consul
This used to be done by /etc/init.d/mp-replyts but that isn't called by
deploy.py, which instead calls daemontools directly. This is part of the
fix for the 502s issue that MP was having.

Let it be known that it is no longer possible to configure the http port
that Comaas is running on using the property 'replyts.http.port'.
Instead, if you want to run on anything but the default port 8080, you
will have to set an environment variable: COMAAS_HTTP_PORT. This was
added earlier already, as part of the preparations for running under
Nomad.
I understand that this means that we cannot use the QA environment as it
currently is anymore, since that requires multiple instances of Comaas
to run on the same host on different ports. QA is currently not being
deployed to, and we should let Nomad find an open port for a new service
instead of hard coding it.
I had to remove 'replyts.http.port' from the properties because of the
way we are loading properties from the various sources. In
CloudDiscoveryConfiguration.initializeDiscovery() we kick off the
application using 'lifecycle.onApplicationEvent(...)' which loads the
various property sources as well. In this call we initialize a
EmbeddedServletContainerInitializedEvent with an overridden getPort()
method. The port must thus be configured already... but this happens
before loading the property sources! Chicken & egg.

This patch needs an accompanying change in Salt that sets a new
environment variable called COMAAS_FABIO_URL_PREFIXES. This should hold
the urlprefixes that are sent to Consul as tags and are then picked up
by Fabio.

As a bonus, I've added hashi-ui as the default Consul web viewer when
running with Docker Compose.

Change-Id: I18ab6b9c10e117982f9ee36abc4cc4ffc35c124b
Reviewed-on: https://gerrit.ecg.so/60700
Tested-by: Gerrit-At-Gitmaster001
Reviewed-by: Jasper van Veghel <jvanveghel@ebay.com>
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Geert Gerritsen <ggerritsen@ebay.com>
Reviewed-by: Gerrit-At-Gitmaster001