COMAAS-505 - use fallthrough retry policy as a default
See the http://docs.datastax.com/en/drivers/java/2.1/com/datastax/driver/core/policies/DefaultRetryPolicy.html
to know what we were using before.

The problem with the DefaultRetryPolicy is that it was causing uneven
response times, holding servlet threads for up to several minutes.
The thing is that no one of our clients waits for so long, so we were
just wasting threads.

Change-Id: I8ff0fd618b2cd71c659113c7ca783e927367a4e3
Reviewed-on: https://gerrit.ecg.so/60873
Tested-by: Gerrit-At-Gitmaster001
Reviewed-by: Jasper van Veghel <jvanveghel@ebay.com>
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Geert Gerritsen <ggerritsen@ebay.com>
Reviewed-by: Gerrit-At-Gitmaster001