Explicitly hint query idempotence
This change is related to enabling speculative execution policy in
COMaaS. A query can only be executed with the speculative policy if
it is known to be idempotent.
When testing speculative execution policy locally, I was using a
query builder in my small test. Turns out, that the query builder
is able to infer a query idempotence (which make sense), while
queries defined as literals (=all comaas queries) require manual
hinting.
Luckily we already have our "modifying" flag, so that I decided to
use "!modifying" condition to explicitly infer idempotence of a query,
which should be safe enough. It may not cover all the cases of writing
queries, but at least we cover all the cases of reading queries, and
that is already good enough to mitigate most of latency spike issues
we have at the moment.

Change-Id: Idc43ca3691b4b0dc8efc65f04968297fbfb1786a
Reviewed-on: https://gerrit.ecg.so/61556
Tested-by: Gerrit-At-Gitmaster001
Reviewed-by: Aleksei Kobiakov <akobiakov@ebay.com>