COMAAS-432 - step 2: refactor and improve the email providing engine
1. DropfolderMessageProvider (aka FilesystemMailDataProvider)
now provides a single message per invocation to unload it from
handling interruption and generify it on the upper level.
2. MessageProcessingService (aka WorkerPoolManager) has been refactored
in favour of executors instead of manual thread group management,
also interruption control has been moved on this level, so that
interruption handling is now responsibility of a single component.

Change-Id: I193641d23b30dc0a29510edc0fd5aa707a69260e
Reviewed-on: https://gerrit.ecg.so/62777
Tested-by: Gerrit-At-Gitmaster001
Reviewed-by: Aleksei Kobiakov <akobiakov@ebay.com>