COMAAS-627 - limit amount of record in a frame for conv. event
Limits an amount of rows of the core_conversation_events put
into a single frame by server at a time.
The limit is set on a per-statement basis.
The SELECT_FROM_CONVERSATION_EVENTS statement is currently unbounded
and may return (and it does) thousands of events.
At the same time, on the driver level there's a limit of 256MB for a
single frame. Exceeding this limit causes an internal driver error
which makes the app unusable until the app is restarted.

In theory, the default value of 100 should cover most of the
conversations, so that they're still fetched in a single round-trip.

Change-Id: I719e6f2c48a45fb2a7901674ccdc1cc75c74a9f6
Reviewed-on: https://gerrit.ecg.so/64826
Tested-by: Gerrit-At-Gitmaster001
Reviewed-by: Remmelt Pit <rpit@ebay.com>
Reviewed-by: Oleksandr Shmatko <oshmatko@ebay.com>
Reviewed-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>
Reviewed-by: Andrea Iacono <aiacono@ebay.com>