Updated smtp endpoint for sending mail
When Comaas is running in a container, postfix does not run on localhost. In the future we will have a postfix cluster, but for now we will connect to the postfixes running on the core nodes instead.

This fixes CS not being able to send mail after it was held.