Fix the issue where pushing to kafka cannot be disabled.
Remove duplication when pushing document updates to kafka

Adding cross DC ES into comaas to be able to do
the index verification.
Comaas should work as before at this point.
Main change is in the ElasticSearchClientConfiguration class.
Only a qualifier was added to other classes.
No actual tests yet.

(while this commit is not related to the previous one
it's still part of the same functinality and I though
you are going to have a hard time merging if I made this
change in other PR)

Adding cross DC ES into comaas to be able to do
the index verification.
Comaas should work as before at this point.
Main change is in the ElasticSearchClientConfiguration class.
Only a qualifier was added to other classes.
No actual tests yet.

(while this commit is not related to the previous one
it's still part of the same functinality and I though
you are going to have a hard time merging if I made this
change in other PR)

Whenever search request comes in, run it against
local and remote (cross-DC) elasticsearch.
If these two properties are found
xdc.search.es.endpoints
xdc.search.es.clustername

Otherwise search funcitonality is unchanged.
Please note that the index name in remote DC should be the same
as in local DC.

Publish total search timer, match and mismatch counter.

Adding bean name to tests to fetch the correct ES Client.