Sanity in the cron job starting properties (#397)
This introduces the active.dc=<dc> property.

Switching cron jobs to the other DC for a failover was unnecessarily
complicated. We had to figure out which cron jobs are enabled, then
change all properties. I made a switch to change all of them at once, so
we cannot forget any. Also we no longer have the ugly ...enabled=ams1
construction. Boolean value enabled can now be true or false.
Downside is that we cannot run some cron jobs in one and some in the
other DC anymore, but that is fine.
Also removed two cron jobs that were only related to riak and also not
enabled in any Bolt tenant.
Also removed a bunch of unused properties.
Also renamed (all?) cronjob enable properties to cronjob.*.