Enable Hazelcast health check, make IP mandatory (#440)
We now have a docker0 interface on the core nodes as well, so we have to
supply the HC driver with an IP to bind to. Make the code reflect this,
so it cannot be read that supplying the IP is not needed.

Also enable the built in HC health endpoint so we can have a health
check in Consul.