Removed dropfolder processing code and core packaging (#766)
We no longer need this.

Not that the
@ConditionalOnProperty(name = "mail.provider.strategy", havingValue = "kafka")
still defaults to false in KafkaMessageProcessingPoolManager.

The allocations that are not newmsg will then not start the PoolManager
at all, reducing memory usage. In comaas.nomad, the newmsg allocations
will still have the value set to 'kafka', like before.

Core nodes can now be switched off and removed.