Adds support for the new tenant: Move Canada (#847)
Adds support for the new tenant: Move Canada

The chosen short tenant name is `mvca`, long tenant name is the same.
No production properties for now (still many things to clarify with KJCA).

Note: no KJCA listeners / eventpublishers are used for MVCA for now. This still needs to be clarified.