Always emit markAsRead events. (#940)
We assessed the markAsRead usage by KJCA and the ratio of calls having
no effect to calls actually zeroing unreads is 3:1.
We also have race conditions related to marking as read due to the events
being not always emitted. For now we decided that it's fine to always
emit the events.